Fix text and fading bugs
Example of the fading bug is narcissu's Translation->Voiced pages at the
beginning should both be visible at the same time during the transition,
but weren't.

This was fixed by changing the SurfaceBlendModes (SDL1.2 didn't have
these I think). Some of them still might be wrong; todo, figure out if
that's the case.

The text bug was also probably fixed by fiddling with those blendmodes.

The flush methods are still in need of cleanup / reexamination