Fix subtitles to display, though inefficiently
Smpeg video subtitles, woo
It's done in a rather hacky way; the video is stuck in a texture, then
rendered to a surface, then blended with subtitles, and then converted
back to a texture.
It would be nice to improve this. Perhaps doing this while moving
to a better subtitle format would make sense.