Rerender frequently via extra ever-present event
This method has a small, but definite, performance impact.
The performance should be evaluated before this gets merged.

Consideration: This could be #ifdef'd for steam, but some people
might run non-steam games via steam and expect the overlay work
well which would be a paint for making it the default, not an ifdef